[gh-#975] move data related services to its own directory
